requirements.txt
alembic==1.11.1
asttokens==2.2.1
backcall==0.2.0
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
decorator==5.1.1
executing==1.2.0
Flask==2.3.2
Flask-JWT-Extended==4.5.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipython==8.12.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib-inline==0.1.6
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
PyJWT==2.8.0
PyMySQL==1.0.3
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.15
stack-data==0.6.2
traitlets==5.9.0
typing-extensions==4.6.2
wcwidth==0.2.6
Werkzeug==2.3.4
zipp==3.15.0

